% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/model.R
\name{aifParameters}
\alias{aifParameters}
\title{Parameters for Arterial Input Functions}
\usage{
aifParameters(type, user = NULL)
}
\arguments{
\item{type}{is one of the following character strings associated with an
AIF: \itemize{ \item\code{tofts.kermode} \item\code{fritz.hansen}
\item\code{orton.exp} \item\code{orton.cos} \item\code{user}
\item\code{empirical} }}

\item{user}{is a vector of estimated AIF parameters or the empirical AIF
values.}
}
\value{
...
}
\description{
Specification of parameters for arterial input functions (AIFs)
}
\details{
...
}
\author{
Brandon Whitcher \email{bjw34032@users.sourceforge.net}
}
\seealso{
\code{\link{compartmentalModel}}, \code{\link{dcemri.lm}}
}
\keyword{misc}

