% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dcemriS4-package.R
\docType{package}
\name{dcemriS4-package}
\alias{dcemriS4}
\alias{dcemriS4-package}
\title{dcemri: A Package for Medical Image Analysis (S4 implementation)}
\description{
A collection of routines and documentation that allows one to perform a
quantitative analysis of dynamic contrast-enhanced or diffusion-weighted MRI
data.  Medical imaging data should be organized using either the Analyze or
NIfTI data formats.
}
\details{
Further information is available in the following vignettes:
\tabular{ll}{ \code{dcemriS4} \tab dcemriS4(source, pdf)\cr }
}
\examples{
\dontrun{
demo(avg152T1)
demo(avg152T1LR)
demo(avg152T1RL)
demo(buckley)
demo(filtered_func_data)
demo(zstat1)
}
}
\author{
Brandon Whitcher \email{bwhitcher@gmail.com}\cr Volker Schmid
  \email{volkerschmid@users.sourceforge.net}\cr Andrew Thornton
  \email{zeripath@users.sourceforge.net}
}
\references{
Schmid, V., Whitcher, B., Padhani, A.R., Taylor, N.J. and Yang,
  G.-Z.  (2006) Bayesian methods for pharmacokinetic models in dynamic
  contrast-enhanced magnetic resonance imaging, \emph{IEEE Transactions on
  Medical Imaging}, \bold{25} (12), 1627-1636.

  Schmid, V., Whitcher, B., Padhani, A.R. and G.-Z. Yang (2009) A
  semi-parametric technique for the quantitative analysis of dynamic
  contrast-enhanced MR images based on Bayesian P-splines, \emph{IEEE
  Transactions on Medical Imaging}, \bold{28} (6), 789-798.
}
\keyword{package}

