\name{expConv}
\alias{expConv}
\title{
  Convolution 
}
\description{
  ...
}
\usage{
expConv(input, k1, k2)
}
\arguments{
  \item{input}{...}
  \item{k1}{...}
  \item{k2}{...}
}
\details{
...
}
\value{
  ...
}
% \references{}
\author{Brandon Whitcher \email{bjw34032@users.sourceforge.net}}
% \seealso{}
% \examples{}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{misc}
