% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_matching.R
\name{findCenter}
\alias{findCenter}
\title{Find the Center of a Binary Mask}
\usage{
findCenter(M)
}
\arguments{
\item{M}{is a binary mask (multidimensional array of logical values).}
}
\value{
A vector of values the same length as the input array.
}
\description{
The center of a binary mask is determined.
}
\details{
This method most likely only works with convex three-dimensional shapes
(e.g., a hyper-rectangle).  Further testing is required to know the limits
of the current implementation.
}
\examples{

M <- array(FALSE, rep(10,3))
M[6:10,6:10,6:10] <- TRUE
Mc <- findCenter(M)
print(Mc)

}
\seealso{
\code{\link{fastTemplateMatching}}
}
\author{
Brandon Whitcher \email{bwhitcher@gmail.com}
}
\keyword{misc}
