% Part of the dcemri package for R
% Distributed under the BSD license: see dcemri/COPYING
%
% $ Id: $

\name{readANALYZE}
\alias{readANALYZE}
\alias{read.analyze.content}
\title{readANALYZE}
\description{
  ...
}
\usage{
}
\arguments{
}
\details{
}
\value{
}
\references{
}
\seealso{
  \code{\link{readNIfTI}}
}
%\examples{}
\author{}
%\keyword{}
