% Part of the dcemri package for R
% Distributed under the BSD license: see dcemri/COPYING
%
% $Id: $

\name{reorient}
\alias{reorient}
\alias{invertReorient}
\title{reorient}
\description{
  ...
}
\usage{
reorient(nim, data, verbose=FALSE, invert=FALSE)
invertReorient(nim, verbose=FALSE)
}
\arguments{
  \item{nim}{}
  \item{data}{}
  \item{verbose}{}
  \item{invert}{}
}
\details{
  ...
}
\value{
  ...
}
%\references{}
%\seealso{}
%\examples{}
\author{Andrew Thornton <\email{zeripath@users.sourceforge.net}>}
%\keyword{}
