% Part of the dcemri package for R
% Distributed under the BSD license: see dcemri/COPYING
%
% $Id: $

\name{translateCoordinate}
\alias{translateCoordinate}
\title{translateCoordinate}
\description{
  ...
}
\usage{
translateCoordinate(i, nim, verbose=FALSE)
}
\arguments{
  \item{i}{}
  \item{nim}{}
  \item{verbose}{}
}
\details{
  ...
}
\value{
  ...
}
%\references{}
%\seealso{}
%\examples{}
\author{Andrew Thornton <\email{zeripath@users.sourceforge.net}>}
%\keyword{}
