% Part of the dcemri package for R
% Distributed under the BSD license: see dcemri/COPYING
%
% $ Id: $

\name{as.nifti}
\alias{as.nifti}
\title{as.nifti}
\description{
  ...
}
\usage{
}
\arguments{
}
\value{
}
\details{
}
%\references{}
%\seealso{}
\examples{
}
\author{Andrew Thornton <\email{zeripath@users.sourceforge.net}>}
%\keyword{}
