% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/model.R
\name{expConv}
\alias{expConv}
\title{Convolution of Exponential Functions}
\usage{
expConv(input, k1, k2)
}
\arguments{
\item{input}{...}

\item{k1}{...}

\item{k2}{...}
}
\value{
The convolved time series.
}
\description{
...
}
\details{
...
}
\author{
Brandon Whitcher \email{bwhitcher@gmail.com}
}
\keyword{misc}

