% Part of the dcemri package for R
% Distributed under the BSD license: see dcemri/COPYING
%
% $Id$

\name{quaternion2rotation}
\alias{quaternion2rotation}
\title{Convert Quaternion into a Rotation Matrix}
\description{
  ...
}
\usage{
quaternion2rotation(b, c, d)
}
\arguments{
  \item{b}{}
  \item{c}{}
  \item{d}{}
}
\details{
  ...
}
\value{
  ...
}
%\references{}
%\seealso{}
%\examples{}
\author{Brandon Whitcher <\email{bjw34032@users.sourceforge.net}>}
%\keyword{}
