% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rcbv.R
\docType{methods}
\name{rCBV.fast}
\alias{rCBV}
\alias{rCBV.fast}
\alias{rCBV.fast,anlz-method}
\alias{rCBV.fast,array-method}
\alias{rCBV.fast,nifti-method}
\title{Regional Cerebral Blood Volume}
\usage{
rCBV(Ct, Ca, time, Hf=1, rho=1) \S4method{rCBV.fastarray}(signal,
mask, aif, time, multicore=FALSE, verbose=FALSE)

\S4method{rCBV.fast}{array}(signal, mask, aif, time, multicore = FALSE,
  verbose = FALSE)

rCBV(Ct, Ca, time, Hf = 1, rho = 1)
}
\arguments{
\item{signal}{is a multidimensional array of signal intensities (or
concentrations).  The last dimension is assumed to be a function of the
acquisition times, while the previous dimenions are assued to be spatial.}

\item{mask}{is a (logical) multidimensional array that identifies the voxels
to be analyzed.}

\item{aif}{Arterial Input Function.}

\item{time}{is the vector of acquisition times associated with the dynamic
data.}

\item{multicore}{is a logical variable (default = \code{FALSE}) that allows
parallel processing via \pkg{multicore}.}

\item{verbose}{is a logical variable (default = \code{FALSE}) that allows
text-based feedback during execution of the function.}

\item{Ct}{...}

\item{Ca}{...}

\item{Hf}{...}

\item{rho}{...}
}
\value{
A \code{nifti} object containing the estimates of regional cerebral
blood volume (rCBV).
}
\description{
This is a set of functions that...
}
\author{
Brandon Whitcher \email{bjw34032@users.sourceforge.net}
}
\keyword{misc}

