% Part of the dcemri package for R
% Distributed under the BSD license: see dcemri/COPYING
%
% $Id: $

\name{readNIfTI}
\alias{readNIfTI}
\alias{read.nifti.content}
\title{readNIfTI}
\description{
  ...
}
\usage{
readNIfTI(fname, verbose=FALSE, warn=-1, reorient=TRUE)
read.nifti.content(fname, onefile=TRUE, gzipped=TRUE, verbose=FALSE,
                   warn=-1, reorient=FALSE)
}
\arguments{
  \item{fname}{}
  \item{verbose}{}
  \item{warn}{}
  \item{reorient}{}
  \item{onefile}{}
  \item{gzipped}{}
}
\details{
}
\value{
}
\references{
  NIfTI-1 \url{http://nifti.nimh.nih.gov/}
}
\seealso{
  \code{\link{readANALYZE}}
}
%\examples{}
\author{Brandon Whitcher, Volker Schmid, Andrew Thornton}
%\keyword{}
