% Part of the dcemri package for R
% Distributed under the BSD license: see dcemri/COPYING
%
% $Id: $

\name{readNIfTI}
\alias{readNIfTI}
\alias{read.nifti.content}
\title{readNIfTI}
\description{
  ...
}
\usage{
}
\arguments{
}
\details{
}
\value{
}
\references{
}
\seealso{
  \code{\link{readANALYZE}}
}
%\examples{}
\author{}
%\keyword{}
