% Part of the dcemri package for R
% Distributed under the BSD license: see dcemri/COPYING
%
% $Id: $

\name{writeNIfTI}
\alias{writeNIfTI}
\title{writeNIfTI}
\description{
  ...
}
\usage{
}
\arguments{
}
\details{
}
\value{
}
\references{
}
\seealso{
  \code{\link{writeANALYZE}}
}
%\examples{}
\author{}
%\keyword{}
